# Fixes #15 -- use cached struct spec to improve performance