Before:

```
alex@penguin ~/p/go-fixedwidth> go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1-4                 	  200000	      7802 ns/op
BenchmarkUnmarshal_MixedData_1000-4              	    2000	    986633 ns/op
BenchmarkUnmarshal_MixedData_100000-4            	      20	  96740143 ns/op
BenchmarkDecode_CodePoints_MixedData_1_Ascii-4   	  100000	     10218 ns/op
BenchmarkDecode_CodePoints_MixedData_1_UTF8-4    	  100000	     11994 ns/op
BenchmarkUnmarshal_String-4                      	 1000000	      1977 ns/op
BenchmarkUnmarshal_StringPtr-4                   	  500000	      2307 ns/op
BenchmarkUnmarshal_Int64-4                       	  500000	      2131 ns/op
BenchmarkUnmarshal_Float64-4                     	  500000	      2193 ns/op
BenchmarkMarshal_MixedData_1-4                   	  200000	      5846 ns/op
BenchmarkMarshal_MixedData_1000-4                	     300	   5799333 ns/op
BenchmarkMarshal_MixedData_100000-4              	       2	 753887899 ns/op
BenchmarkMarshal_String-4                        	  500000	      2666 ns/op
BenchmarkMarshal_StringPtr-4                     	  500000	      2692 ns/op
BenchmarkMarshal_Int64-4                         	  500000	      2327 ns/op
BenchmarkMarshal_Float64-4                       	  500000	      2882 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	27.081s
```

After:

```
alex@penguin ~/p/go-fixedwidth> go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1-4                 	  300000	      5556 ns/op
BenchmarkUnmarshal_MixedData_1000-4              	    3000	    877359 ns/op
BenchmarkUnmarshal_MixedData_100000-4            	      20	  66129486 ns/op
BenchmarkDecode_CodePoints_MixedData_1_Ascii-4   	  200000	      8854 ns/op
BenchmarkDecode_CodePoints_MixedData_1_UTF8-4    	  200000	     10834 ns/op
BenchmarkUnmarshal_String-4                      	 1000000	      1833 ns/op
BenchmarkUnmarshal_StringPtr-4                   	 1000000	      2049 ns/op
BenchmarkUnmarshal_Int64-4                       	 1000000	      1901 ns/op
BenchmarkUnmarshal_Float64-4                     	 1000000	      2057 ns/op
BenchmarkMarshal_MixedData_1-4                   	  200000	      5659 ns/op
BenchmarkMarshal_MixedData_1000-4                	     300	   4849681 ns/op
BenchmarkMarshal_MixedData_100000-4              	       2	 799881256 ns/op
BenchmarkMarshal_String-4                        	  200000	      7547 ns/op
BenchmarkMarshal_StringPtr-4                     	  200000	      5303 ns/op
BenchmarkMarshal_Int64-4                         	  300000	      3732 ns/op
BenchmarkMarshal_Float64-4                       	  300000	      3968 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	28.350s
```